Add OpenJDK 8 add-on cartridge.
This add-on cartridge sets an environment variable called
OPENSHIFT_OPENJDK8_JAVA_HOME which can be consumed by Java web_framework
cartridges.

The precise name is "java-openjdk-1.8.0". It may be added to any Java
web_framework cartridge. Though, that cartridge needs to support Java 8,
which does not seem to be the case for the current jbossas cartridge.
The hosting cartridge will have to have support for this add-on
cartridge as well. Steps for adding support are:

1. Check for a java8 marker file
2. If a java8 marker file exists use OPENSHIFT_OPENJDK8_JAVA_HOME
   for setting JAVA_HOME

Example usage with a modified jbossews-2.0 cartridge:

$ rhc app create testapp jbossews-2.0 openjdk8

Then in order to switch to Java 8 use the following:

$ cd testapp
$ git mv .openshift/markers/java{7,8}
$ git commit -m "Using Java 8"
$ git push origin master

Testing done:

I've verified this cartridge works with a modified jbossews-1.0 and
jbossews-2.0 cartridge. The spec file builds successfully. The resulting
RPM installs fine and the RPM-installed cartridge works also with
jbossews-1.0 and jbossews-2.0.