broker: add PREVENT_ALIAS_COLLISION option.
https://trello.com/c/ASq1CXyv
Normally when an alias is created, OpenShift validates that it is not
already taken and that it is not of a form that could conflict with an
app name; this prevents problems where the node frontend could have two
destinations for the same URL on the same node and can only route one.
Normally it is obviously desirable to prevent this. But in certain use
cases we want to relax this validation, so this option does that (when
set to false). It is not documented in broker.conf comments because it
should not be used unless you really, really know what you're doing.