cloud_user: enable normalization of user logins.
https://trello.com/c/noKWhxNG/212-2-enable-custom-normalization-of-usernames-from-external-sources

This enables an administrator to define a custom method to normalize
usernames before they are translated into cloud_user entries.

For example, this method could downcase the login such that the login
is case-insensitive; then case-insensitive authentication methods would
not create multiple cloud_users for the "same" login just because the
user capitalized it on one occasion and not another.