oo-diagnostics: Add test_node_clock_in_synch_with_broker
Add a test_node_clock_in_synch_with_broker check to oo-diagnostics.  This
test is specific to node hosts.  It sends an HTTP request to the host
identified by the BROKER_HOST setting in /etc/openshift/node.conf and
compares the date in the Date: header of the response to the date of the
node host.  If the difference is equal to or more than 55 seconds, the
check gives an error.  Else, if the difference is equal to or more than
5 seconds, the check gives a warning.

This commit fixes bug 1121267.