oo-accept-systems: Drop obsolete/downloaded carts
Fix the check_nodes_cartridges test in oo-accept-systems to ignore obsolete
cartridges returned by the node hosts, because the broker will not list
obsolete cartridges, and to ignore downloaded cartridges returned by the by
the broker, because a node will not have downloaded cartridges in its
cartridge repository.

This commit fixes bug 1124938.