The output message about TLS URLs is not clear when creating jboss-amq cartridge
https://bugzilla.redhat.com/show_bug.cgi?id=1128083

Jboss-AMQ provides 4 ways to connect. They are openwire, stomp,amqp and
mqtt. But when creating jboss-amq app, the output message is not clear
which one is for stomp, which one is for amqp.
There's no information related with these ports in app's gear env
variables. (Who will generate these variables?  Openshift Framework or
Cartridge)
User only have one way to guess which port is for openwire. 'TLS_PORT_1'
which is defined in manifest.yml is for "OPENWIRE_SSL_PROXY_PORT". It
will use the smallest port 2303. And 'TLS_PORT_4' uses the biggest port
2306.