node conf: supply resource_limit.conf examples
Bug 1133936 - Provide some usable canned resource_limits profiles
https://bugzilla.redhat.com/show_bug.cgi?id=1133936

Additionally, reorganizes resource_limits.conf a little bit to make it a
little easier for an administrator to find common things they want to
change.

Finally, modified resource_limits.conf defaults to be in line with
a small m3.xlarge host. The CPU constraints are relaxed a bit and
max_active_gears is lowered to be more conservative. A README is added
to indicate that customers are expected to customize this.