broker: ensure normalization is idempotent
https://bugzilla.redhat.com/show_bug.cgi?id=1131317
Bug 1131317 - Fail to run 'rhc setup' when normalizing user with complex matrix on broker

When the username is normalized, then stored, then retrieved and used in
a new request, it is normalized again. In order for this to match,
normalization should be idempotent. It's possible to come up with
reasonable normalizations that accidentally break this rule. So, to make
sure, we now normalize names repeatedly until the answer stays the same.