Fix scalable platform snapshot test
Add a wait in between when the mock-plugin cartridge is added and when
the snapshot command is issued. The addition of the plugin causes
haproxy to reload, but it does it gracefully, which means that the
previous haproxy process instance may still be running while a new one
is running. Attempts to rapidly add a cartridge and then snapshot the
application (which stops haproxy, does some snapshot work, and then
starts haproxy again) currently may fail due to the way the haproxy
cartridge's control script checks to see if haproxy is running.