oo-accept-systems: fix errors from PR 5851
The cartridge consistency checks introduced in PR 5851 did not correctly
determine which cartridges were active on the broker, which becomes
apparent once you have imported more than one version of any cartridge.
Inactive versions of the cartridge could be compared against the node
versions. This is now fixed.

Also obsoletes were added to the list of things not to compare, and
probably some other issues addressed along the way.