BZ#1128857 - Fixes failover of activemq hosts during broker publish
* We were previously reusing a Stomp::Connection which is normally fine
and actually preferred. However, in the failover scenario the connection
that the stomp library has validated goes away, which in certain
failover scenarios causes the first message in the sequence of messages
associated with a routing event to be missing. By reinitializing the
connection when we send a message, the missing message scenario is
avoided.

* Also removing reliable => true in the connection opening as it is implicit.