node: fix secondary haproxy app fqdn
The secondary HAproxy gear in a HA application should be accessible at
the node httpd proxy via the app fqdn. Previously this extra route was
added as a sort of additional unofficial fqdn for the gear frontend that
resulted in some inconsistencies. This change updates the method used
such that the app fqdn is added as an alias on the gear frontend instead.

Bug 1155677 - HA apps add fqdn entries in frontend http configuration
causing oo-accept-node to FAIL
https://bugzilla.redhat.com/show_bug.cgi?id=1155677