Add F5 HTTPS support and removing the routes feature
- This patch adds the ability to have custome alias https termination
at the F5 LTM. This requires SNI configuration in F5. You can refer to
https://support.f5.com/kb/en-us/solutions/public/13000/400/sol13452.html
for details on how to set that up. You also still need to set up an HTTP
virtual server and HTTPS virtual server. The fallback client ssl profile
needs to be manually added to the HTTPS virtual server. It's important
to set the fallback client ssl profile up with the node https cert/key
pair and mark the client ssl profile as the default SNI profile.

- Since this patch requires that cert/key files are copied to the F5 LTM
as well as removed when they are no longer required, there has to be a
manual configuration of the LTM admin user's authorized_keys to allow
ssh access to the LTM from the routing daemon host. Additionally, a
manual setting of the admin user's terminal type to "Advanced shell" is
required to execute remote cert/key removal commands.