oo-admin-ctl-routing: Fix delete-monitor help text
Fix the help text for oo-admin-ctl-routing's delete-monitor command.

The delete-monitor command takes arguments <name> <type> [<pool>].  It
makes sense for the monitor name and type to be together and for the
optional argument to be last, so rather than changing ordering of arguments
in the command, it makes more sense to change the help text to match the
way the command actually works.

This commit is related to bug 1199904.