routing-daemon: Refresh monitors in case of error
Modify the controllers' create_monitor and delete_monitor methods to
refresh the controller's cached list of monitors from the load balancer
when the controller is told to create a monitor that already exists in
the controller's list, or if it is told to delete a monitor that does
not exist in its list.

After this commit, the routing daemon is better able to cope with the
load-balancer's state's being modified from outside the routing daemon
while the routing daemon is running.

This commit is related to bug 1199904.