routing-daemon: Try harder to create pool
Modify OpenShift::RoutingDaemon#create_application to rescue and log any
LBControllerException raised in the controller's create_monitor method
so that create_application will continue on to try to create the pool.

The goal of this commit is to make the routing daemon try a little harder
to cope with inconsistent state.

This commit is related to bug 1199904.