Install cpan dependencies in listed order
Respect the order in which dependencies are listed in .openshift/cpan.txt
or deplist.txt, and install them prior to detected dependencies.

Uniqueness is still enforced by excluding subsequent duplicates.

An instance where ordering matters is with Dancer and Plack::Builder.  Even
though installing Dancer causes Plack::Builder to be installed, the
installation of Dancer fails unless Plack::Builder is already installed.
Thus it is required either to retry the installation of Dancer or to
explicitly install Plack::Builder first (which can be done by listing it
before Dancer in .openshift/cpan.txt).

This commit fixes bug 1223986.