routing-daemon: F5: check for and log SSH errors
F5IControlRestLoadBalancerModel: Add run method that runs the given
command, captures stdout and stderr, and raises an exception if the
command fails.

F5IControlRestLoadBalancerModel#add_ssl: Use the new run method to check
for, log, and handle errors from the ssh and scp commands.  Modify the
logging output to be neater and more consistent.

This commit fixes bug 1226061.