routing-daemon: controllers: invoke model update
AsyncLoadBalancerController#update, BatchedLoadBalancerController#update:
Invoke the model's update method if it exists.

Note that SimpleLoadBalancerController does not define its own update
method, so it inherits the update method from LoadBalancerController,
which also invokes the model's update method.

Note too that the only model that has an update method is
NginxLoadBalancerModel, and we use SimpleLoadBalancerController with
this model, so we haven't been missing the update method in the other
controllers.