Resolve race condition where cartridges activated within the same second conflicted
Bug 1191283
Bugzilla link https://bugzilla.redhat.com/show_bug.cgi?id=1191283

Using the ":priority.lt => self.priority" would not find types with a priority within the same second of this types priority. Because of this, older types did not have their priority properly set to 'nil'. This caused us to believe that a cartridge type with the same name was activated at the same time (or after) this type.