Improve error reporting for member add/remove/update through oo-admin-ctl-domain
When an error occurs in domain.save, an exception is not returned. Instead, the method does not return true to indicate a failure. These failures were not being caught previously.

Additionally, errors encountered in domain.run_jobs did not populate domain.errors, resulting in an empty hash being reported when an error does occur.