oo-admin-ctl-app: Fix remove-gear ignores min scale setting
Currently, the minimum gear scale limit is ignored when the command
'oo-admin-ctl-app -c remove-gear' is executed causing the application
becomes unscaled when number of gears reaches 1.

The commit will fix a check to compare the number of gears with the
gear limit  which is retrieved from group_overrides. If the number
of gear reaches gear limit, the gear cannot be removed and an error
will be issued. Otherwise, the gear is removed as usual.

Bug <1272195>
Link <https://bugzilla.redhat.com/show_bug.cgi?id=1272195>

Signed-off-by: Vu Dinh <vdinh@redhat.com>