Bug 1064039: Add oo-diagnostic report 401 Unauthorized error
The 'oo-diagnostics' currently won't fail if there are 410 Unauthorized
error in the production/development.log. After this change, 'oo-diagnostics'
will check either production or development.log depending on which mode the
broker is in. If the 401 errors exist, an error message will be displayed
with the sugguested solution to fix the issue. The error message may
continue to be displayed in the subsequential 'oo-diagnostics'
until the 401 error message is cleared up in the next log rotation.

Bug <1064039>
Link <https://bugzilla.redhat.com/show_bug.cgi?id=1064039>

Signed-off-by: Vu Dinh <vdinh@redhat.com>