FrontendHttpServer: Recover from missing manifest
OpenShift::Runtime::FrontendHttpServer#initialize: Rescue any exception
from initializing the cartridge model, and set @standalone_web_proxy to
false in that contingency.

Before this commit, failure to initialize the cartridge model would cause
a failure to initialize the frontend http server, which would cause
a failure to initialize the container plugin, which would prevent the
container plugin's destroy method from finishing.  Consequently, it was
impossible to delete a gear with a bad manifest.yml file.

This commit fixes bug 1273658.