Bug 1264722: Fix oo-register-dns shows erros with any option
There is a command conflict as '-h' option is linked link both '--help'
and '--with-node-hostname' which causes 'getoptlong' object to fail.
In fact, '--help' should be linked to '-?' instead. This commit recrifies
the issue by linking '--help' back to '-?' as it shows in help message.

Bug <1264722>
Link <https://bugzilla.redhat.com/show_bug.cgi?id=1264722>

Signed-off-by: Vu Dinh <vdinh@redhat.com>