Fix gear extended tests
When appending a line for the GEAR_UPGRADE_EXTENSION setting to
/etc/openshift/node.conf, include a leading '\n' EOL marker in case
the last line of node.conf is incomplete and is missing the EOL marker.
Otherwise the GEAR_UPGRADE_EXTENSION setting will be appended to the
incomplete line and will not take effect, and as a consequence, the test
will fail the first time it runs (although it will succeed on subsequent
runs).

Change the clean-up code to delete the entire line for the
GEAR_UPGRADE_EXTENSION setting instead of substituting an empty line,
and then delete the last line if it is empty.