routing-daemon/README.md: Various fixes
Fix a grammatical error in the "Configuring the Daemon" section.

Delete `BIGIP_MONITOR` from the list of settings for the F5 BIG-IP LTM
controller, and add `BIGIP_SSHKEY` and `BIGIP_DEVICE_GROUP`.

Make it clear that the HTTP virtual server and the HTTPS virtual server may
share a vip.

Reorganise the documentation for F5 BIG-IP LTM for readability.

Document the `BIGIP_DEVICE_GROUP` setting.

Correct the description of the local-traffic policy is created to say that
the daemon creates it on initialisation and adds and removes rules during
operation, and mention application aliases and SSL certificates and keys as
well.

Change "bash" to the more generic "shell".

Delete mention of an old default template for pool names
("/Common/ose-#{app_name}-#{namespace}") from an earlier, never-shipped
version of the F5 backend, and instead refer to the section "Pool and Route
Names", which mentions the current default template.

Specify that F5 BIG-IP LTM pools contain proxy gears, as opposed to
application gears (unless overridden by the `ENDPOINT_TYPES` setting).

Add `NGINX_SSL_CERTIFICATE`, `NGINX_SSL_KEY`, `HTTP_PORT`, and `SSL_PORT`
to the list of settings for the nginx controller.

Mention the nginx configuration files for application aliases and SSL
configuration.

Document the `HTTP_PORT` and `SSL_PORT` settings.

Document the `ENDPOINT_TYPES` setting.