deployments_controller.rb: Adds check for REST deployment without arguments
Bug 1223025
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1223025

Currently, attempting to change a deployment and not passing in a target variable (either ref or artifact_url) will result in a false success message.

Adds a check to ensure that a ref or artifact_url variable is present before attempting to deploy.  Also adds tests for the scenario where neither variable is present and the scenario for when both variables are present.