Bug 1170540 - False success oo-admin-cartridge --recursive/-R in invalid dir
When a cartridge is installed with recursive option and an invalid directory,
the oo-admin-cartridge command fails to report error and falsely succeed.

This commit adds a invalid/empty directory check to the oo-admin-cartridge
to stop execution and report the error to the user.

Bug 1170540
Link <https://bugzilla.redhat.com/show_bug.cgi?id=1170540>

Signed-off-by: Vu Dinh <vdinh@redhat.com>