oo-admin-chk: Adds solutions to a couple of error messages
Bug 1111598
BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1111598

Changes the incorrect error message for when a gear exists on the node, but not in mongo.  The error message now tells the user the correct solution for the error.

Also adds a solution to the error message for when there is a mismatch between consumed gears and actual gears.