deployments.rb: Changes :err to be visible to user
Using the REST api for binary deployments could result in errors with the node using the curl command, however the user would not be able to see why or how the curl command failed.

Changes the :out and :err streams sent to the node to be nil, so that the error output is sent back to the user as part of the text variable.  Also changes the corresponding test.

Bug 1126836
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1126836