oo-accept-node: check for missing gear directories
Bug 1281912
BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1281912

There was no check for the situation where a node has a user but is missing the
corresponding gear directory.  The oo-accept-node script checks for valid file
structures, but does not check that each user has a directory.  This change will
cause oo-accept-node to fail when there are users without matching directories
and let the users know that the file is missing.