Bug 1160934 - "oo-admin-ctl-gears stopgear" failed to stop idled gear
During the 'idlegear' operation, the gear in fact is stopped but the
state value is changed to 'idle'. Then, subsequently, when 'stopgear'
operation is excuted, the cartridge doesn't perform any extra works
as the gear is already stopped. However, the state is not modified
to 'stopped' due to lacking of explicit state change in application
container itself.

This commit will add explicit state change to application container
to ensuring the state is changed accordingly.

Bug 1160934
Link <https://bugzilla.redhat.com/show_bug.cgi?id=1160934>

Signed-off-by: Vu Dinh <vdinh@redhat.com>