Bug 1296943: Quickstart deployment via git fails to bind address
When quickstart is deployed with rhc --from-code, there is a start process
(from control script) called during post_receive step which starts the gear and
apache processes. However, in the next step, the cartridge is started as well.
As a result, there are two consecutive start processes leading to the second one
fails due to the fact apache (httpd) process is already started and currently
in use.

This commit adds a check of current state of the gear before starting the cartridge.
If the gear is already started, the restart cartridge process is triggered instead.
Otherwise, the start process is striggered as expected. This fix only change the
behavior initial build of cartridge during git deployment and it's operating within
the local gear. Therefore, it doesn't impact any other gears that are running in
the same node.

Bug 1296943
Link <https://bugzilla.redhat.com/show_bug.cgi?id=1296943>

Signed-off-by: Vu Dinh <vdinh@redhat.com>