Bug 1299095 - oo-diagnostic test_broker_certificate fails on broker
The 'test_broker_certificate' method in oo-diagnostics fails if it's run
on broker due to the fact the test will checking for a file that only
exists on node but not broker.

This is a regression issue as in PR 6275, several lines of code was added
which was intended to run on the node only. As a result, when the code was
run in the broker, the failure occurs. This commit will separate the code
for broker and node test into 2 separate tests: test_broker_certificate
and test_node_certificate.

Bug 1299095
Link <https://bugzilla.redhat.com/show_bug.cgi?id=1299095>

Signed-off-by: Vu Dinh <vdinh@redhat.com>