Bug 1305688 - oo-accept-broker incorrectly parses MONGO_HOST_PORT
The validation of mongo replicant fails as the code in oo-accept-broker
that is responsible to parse MONGO_HOST_PORT is not working correctly
as it fails to remove unwanted quote characters in the variable.

This commit fixes the code that is parsing the MONGO_HOST_PORT to remove
the quote characters before performing the task. As a result, individual
ports and hosts from MONGO_HOST_PORT are now parsed correctly.

Bug 1305688
Link <https://bugzilla.redhat.com/show_bug.cgi?id=1305688>

Signed-off-by: Vu Dinh <vdinh@redhat.com>