Remove double warning msg when adding cartridges
bz1169690
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1169690

When adding cartridges to an app that is approaching quota limits,
the "Warning: Gear <uuid> is using <percent> of disk quota" is printed twice.
This PR removes the report_quota when installing a cartridge, so the
quota is only reported once after the cartridge is installed.